+++ SeaBattleTrophy/SeaBattleTrophy/ViewModels/SeaBattleTrophyGameViewModel.cs
@@ -22,10 +22,10 @@ namespace SeaBattleTrophy.WPF.ViewModels
         {
             _game = new SeaBattleTrophyGameManager();
             SelectedShip = new ValueHolder<IShipReadOnly>(_game.Ships.First());
-            SeaMapSizeInPixels = new ValueHolder<float>(1000);
+            SeaMapSizeInPixels = new ValueHolder<double>(1000);
         }
 
-        public IValueHolder<float> SeaMapSizeInPixels { get; private set; }
+        public IValueHolder<double> SeaMapSizeInPixels { get; private set; }
         
         public IValueHolder<IShipReadOnly> SelectedShip { get; private set; }
